hands_on_microservices/frontend/requirements.txt
# streamlit==1.20.0
# requests==2.28.2
# Pillow==9.5.0
# streamlit-drawable-canvas==0.9.2

streamlit
requests
Pillow
streamlit-drawable-canvas
streamlit_option_menu